refactor(caa upload): use logger instead of banner
Switch to using the newly-introduced logger to log messages, and
use the status banner as a sink for that logger to actually display
the messages. Since we can now also distinguish between the types of
messages, the status banner will use different text colours depending
on the message severity.

We're currently still setting up the page's UI in the `ImageImporter`
constructor, so we're temporarily passing the status banner as an
argument to the constructor to then also provide it to `setupPage`.
We'll refactor this in a future commit.