build: temporarily enable `skipLibCheck` in TS
With the inclusion of `rollup-plugin-consts`, type checking fails
with an error originating from a .d.ts file provided by that plugin.
We can't really do anything about this, except for ignoring errors
in .d.ts files entirely, through the `skipLibCheck` option. We should
re-enable this as soon as the issue is fixed upstream though, as this
option will ignore errors in ALL .d.ts files, including our own.