refactor(caa upload): provide Amazon favicon through getter
Turn the Amazon provider's favicon into a getter instead of a simple
property which gets initialised at instance creation time. Because
the Amazon provider uses `GM_getResourceURL` to retrieve the favicon
to bypass Firefox' Enhanced Tracking Protection, testing anything
which instantiates the Amazon provider required complicated stubbing
of the GM function. Turning it into a getter provides the benefit
that we do not have to stub the function unless we actually use
the favicon, and significantly simplifies testing.