refactor(caa upload): improve testability of form filling
Previously, there was no way to wait until the form filling completed
for all images. This poses a problem during tests, as there, we need
to wait until the form is filled before we can check whether it's
filled correctly. Now, the form filling returns a promise which resolves
whenever all images are fully enqueued. We're also awaiting this now
in the main script, to prevent erroneously filling the edit note when
the enqueuing failed, and to be able to report the error to the user.