feat(caa upload): allow providers to reject HTTP redirects
Providers may unexpectedly redirect to other releases, in which case
the extracted cover art may be incorrect. We now force each provider
to expose a unique ID for a release. When the provider page is fetched
and a redirect was encountered, the IDs for the original URL and
redirected URL is compared to determine whether the redirect was safe.
If it wasn't, an error is thrown and the script will reject extracting
images from the redirected release.