feat(caa upload): improve edit note format
Swaps the order of the original and maximised URLs in the edit note
in preparation for putting the redirected URL in the edit note as
well. The new edit note format is now as follows:

* <original URL as entered by the user/extracted from the provider>
  -> Maximised to <maximised URL>

This makes it clearer which of the URLs are original and which are
obtained after various steps taken by the script.