perf(caa upload): cache Discogs API responses
We were contacting the API every time we maximise an image. This
becomes wasteful as soon as multiple images from the same release
are maximised. To make matters worse, we were also contacting the
API to extract images from a release. For a release with 30 images,
this would lead to 31 identical requests (one for the release itself,
30 for each individual image to be maximised).

Instead, we now use a cache and reuse previous API responses. The
cache is isolated to a single script load (i.e., not shared across
multiple instances, not persistent), so there is no eviction policy.
However, failed requests are not stored in the cache so that they
can still be retried without reloading the page.