ci: enable true continuous deployment
Changing the previous CD setup to remove the requirement to manually
bump the version whenever a release needs to be made. `commit_release`
will now compare the previously built version to a built version of the
most recent code. When any change is detected to the built code, it will
automatically increment the version and commit the new version to `dist`,
as before.