feat(caa upload): extract images from Amazon's embedded JavaScript
Using the thumbnails which are shown on the page has two disadvantages:
- Pages show only a limited number of thumbnails, not all of them.
- Maximised versions of these might still not be the largest available.

Unfortunately the most reliable way (so far) to get all images in their
highest resolution, is to extract an object from the page's embedded
JavaScript. The advantage of this approach is that we can also extract
artwork types from there now (at least 'Front' and 'Back', maybe others).