ci: run deploy on pushes to `main`
We were previously running deploy whenever a PR was merged, so that
we could retrieve PR information. However, since the deploy job first
builds and tests userscripts, and there's also a workflow running on
`main` which builds and tests userscripts, all of this workload was
duplicated. It turns out we can retrieve PR information directly from
pushes to `main`, so we'll now run the deployment on those instead.