ci: make CodeQL workflow dispatchable
For some reason, security alerts aren't getting closed, even though
they are fixed. I suspect it might be because the alerts were first
picked up by a workflow called "CodeQL", but the fixes were noticed
by the "CI" workflow. To try to confirm this suspicion, I'm making
this one dispatchable so that I can manually run it to see if the
alerts close now.