fix(lib): set base URL when parsing DOM
If we don't set the base URL, and the HTML doesn't define its own
base URL, relative URLs will be resolved against the location on
which the HTML was parsed. That's not always what we want. To avoid
any problems with this in the future, we're forcing the base URL to
be set at parsing time.