refactor(caa upload): allow wildcards in provider domains
For providers such as Bandcamp, whose URLs can have arbitrary
subdomains, the previous approach was not sufficient. For Bandcamp
specifically, we added an edge case to the domain extraction logic
of the provider dispatcher, which was neither elegant nor extensible.
Given that we'll need to support more arbitrary subdomains for the
7digital provider (which uses regions as subdomains, e.g.
`de.7digital.com`), a better solution is required.

One option would be to change the providers to expose a function like
`supportsDomain`, which can then be overridden. However, this requires
changes to all providers, and also means we won't be able to efficiently
dispatch to the correct provider. We'd need to perform a linear
traversal of the provider list and run its `supportsDomain` function.
If any of those providers have a buggy implementation, it would affect
all others too.

Instead, we'll allow simple wildcards at the start of a domain pattern.
These wildcards are limited in where they can go, similar to `@match`
stanzas in userscript headers. Wildcards are only allowed at the start
of a domain, and only for subdomains:

Allowed:
- `*.example.com`
- `*.b.example.com`

Not allowed:
- `*.com`
- `amazon.*` (we don't want to match `amazon.fake`)
- `a*.example.com`
- `a.*.example.com`

A pattern such as `*.example.com` will match `x.example.com`,
`x.y.example.com`, as well as `example.com` itself. However, the
pattern `example.com` will *not* match `x.example.com`.

To implement this, we use a tree that maps domain name parts to
leafs (the object to dispatch to for that domain name) or internal
nodes (pointing to subtrees). Each node key is either:
- one part of the domain name (e.g. `x` for `x.example.com`).
  Top-level keys use the "base" domain name, such as `example.com`
  to avoid unnecessary indirection for TLDs.
- the wildcard `*`, to match any remaining domain name parts. This
  is always a leaf node.
- the literal empty string, always pointing to a leaf node. This is
  used when there are multiple patterns with the same prefix, e.g.
  `open.qobuz.com` and `qobuz.com`. This will result in an internal
  node with two keys: `''` pointing to the object for `qobuz.com`,
  and `'open'` pointing to the subtree for `open.qobuz.com` (which
  in this case will also be a leaf object).

The provider dispatcher has been refactored to make use of this new
structure, and the Bandcamp provider's domain has been changed to
use a wildcard.