test(caa upload): use shared spec in URL testing
All of the URL tests are basically the same except for the input
data. This is repetitive and bad for maintainability. I've
refactored the tests so that they only define the test data, and
use a shared spec to actually carry out the test. This will make it
easier to add other tests for the URLs.