feat(caa upload): add VGMdb seeder to workaround private images
The VGMdb provider cannot grab all images for some releases, since
some images are only available to logged in users. As a workaround,
we'll provide a seeder which runs on the VGMdb page if the user is
logged in, which can be used to send cover art to the linked release.