refactor(caa upload): invert control between UI and App
Previously, App called into UI, now, UI will call into App. This
will make it somewhat easier to allow the user to configure certain
parameters (such as fetching only the first image) through the UI.
Without this change, the app would need to provide a callback to the
UI to be fired whenever the configuration state changes, now, the UI
can call the appropriate methods itself.

The main user-visible change is that this will now lead to the input
being cleared regardless of whether the queueing was successful.
However, it makes sense to clear the input automatically even on
failure to allow trying again with the same or a different URL.
Moreover, this fixes an issue where input wasn't cleared when
multiple URLs were entered, separated by whitespace.

We're also moving the App class to a separate file, mainly because
otherwise we need to export it from `index.ts` which leads to
unnecessary warnings when building. It also cleans up the imports
in `index.ts` as an added benefit.