fix(caa upload): improve error for removed Soundcloud releases
Soundcloud is no longer returning a 404 for missing tracks or sets,
and instead returns a 200. Previously, this led to a "Could not extract
metadata" error, which isn't the most descriptive. Now, we indicate
in that error message that the release may have been removed, so that
it doesn't look like an internal error.