fix(caa upload): enqueue images from content context
Greasemonkey scripts that use granted APIs run in the content context,
not the page context, and therefore have no access to the jQuery object
defined by MB. We "escape" this through `unsafeWindow` and `cloneInto`
to trigger the event. We also trigger the handler manually rather than
triggering the event via jQuery, since jQuery tries to call some
functions defined on the event but the page scripts don't have that
permission.