fix(caa upload): grab non-public VGMdb images if user is logged in
The GM.xmlHttpRequest function sends cookies by default on all
major userscript engines, even for cross-origin requests, so if the
user is logged in and we request the page content through GM.xhr,
we'll get a logged in page and can extract all images from there.

We'll still keep the seeder as it may come in useful to some.