build!: update browser support
There's no use in supporting IE since I doubt anyone uses it to run
userscripts (is it even possible?). Also dropping support for old
Safari versions as they cause bloat in the generated code (class
transpilation, for example) and are fairly outdated. Again, probably
nobody uses Safari to run userscripts anyway, but there are userscript
managers for recent Safari versions, so we'll keep transpiling to
that (dropping support doesn't cause a change in transpilation anyway).

We won't be able to test with Safari though, since it we cannot
automatically add extensions at this time.