test(e2e): self-host an MB instance for E2E testing
We're self-hosting an instance of MB during testing to avoid
possible instability on test.musicbrainz.org, and because test.MB
uses an old MBS version. Moreover, individual test scenarios could
run quicker now that the MB instance is hosted on the same machine.

MB docker images are automatically built every week in a separate
repository to ensure we're working with recent MB versions, and to
spare our CI the need to build the images again and again on every
run.