build: split into multiple TS projects
This allows us to separate the different source trees and make sure
that certain projects cannot use things from other projects. E.g.,
now we cannot import the tests in the scripts, which would not make
sense. More importantly, this allows us to selectively include
global ambient declarations for certain projects, but not others.
For example, we include jest's global declarations in the unit tests
by default, but not in the scripts. Because of this, we also won't
get nonsensical autocomplete suggestions in those other projects.
Most importantly, this will prevent type declaration conflicts
between unit and E2E tests, which both use implicitly global
`describe` and `it`, but with different signatures.