fix(work codes): support translated agency IDs
In some languages, the agency code keys are translated leading to
issues pasting the codes. For example, in German, the key for ASCAP
is "ASCAP-ID" while the script expects the English "ASCAP ID". This
is a temporary fix to normalise such translations. Some exceptions,
most notably PRS, may still not work correctly in translations. We
need a better fix for the script to work in other languages, but for
that, some parts would need quite invasive changes.