refactor(caa dims): split cache for dimensions and file info
These two pieces of information are fetched separately, and either
of them can fail. Previously, we took an all-or-nothing approach to
caching. If either of them failed, the other one would not be cached
either. We're now splitting the cache to store both pieces of information
separately, so we'll only store the results which we know to have been
successful.