refactor(caa dims): modularise getting image info
The idea with this modularisation is that we'll be able to create
new subclasses for custom behaviour on how to load image info. For
example, in ECAU, we could create a subclass for a-tisket images,
whose info is loaded in a different way than for CAA. We could
possibly also use a factory somewhere.

TODO: Test.