ci: set up semantic commits checker
I've enabled an app that checks all PR titles and commit messages
for adherence to conventional commits. It should already be running,
but let's tweak the configuration a bit.

The result is that the app will complain on PRs where any of the commits,
and more importantly the title, is not following the CC standard. The
PR title is most important, as that is what we're using to auto-generate
the changelog. These checks aren't marked as required checks, so the PR
can still be merged when they fail, but it serves as a gentle reminder.