fix(ecau): retry when failed to load dimensions on a-tisket
Refactored the image dimensions on a-tisket to use CAA Dims' implementation.
This gives us retrying for free. We're also now using a local storage cache
so that we don't make unnecessary requests if we already know the image
information. The cache is deliberately kept small. We're not using the
IndexedDB cache like in CAA Dims since that'll inflate the size of the
built scripts by a lot, and is probably overkill.