feat(ecau): block edit submissions while processing URLs
While a URL is being processed and added to the form, the edit submission
button will be disabled until processing finishes. This prevents the user
from being able to accidentally submit the edits while the images haven't
been queued fully yet.

Currently, since we're only queueing images and filling the edit notes
after a full batch has been fetched, such situations should occur rarely.
It's still possible if the user is adding multiple URLs though. However,
in the future, we'll switch to queueing individual images as soon as
they're fetched, and this submission blocking will then be a necessity.