fix(split links): submit all but last URL immediately
Previously, we patched the editor to show only the first URL, and
submit the rest of the links when the first one is submitted. This
also led to problems when the user changed the first URL, as the
remaining URLs would be forgotten.

Instead, we'll submit all but the last URLs immediately when they're
pasted, and leave the last URL as pending. This provides a better UX.