refactor: restrict usage of `it` identifier
We shouldn't use `it` as a variable name to describe an iterator,
since that conflicts with the `it` function used in tests. This also
leads to buggy behaviour in the new eslint-plugin-jest max-expects
rule. ESLint would've warned about these usages anyway, but it doesn't
know of the existence of the `it` function in the outer scope.