fix(ecau): indicate progress when fetching single images
Moving the direct image fetching log message to the caller instead
of the fetcher since the caller knows how many URLs need to be processed
whereas the fetcher doesn't. It'd make a bit more sense to have it in
the fetcher, but that'd require more refactoring that I don't want to
do right now.