fix(ecau): grab all Soundcloud track images
Soundcloud pages only contain data for the first 5 tracks in a set,
and the remainder of the tracks need to be queried separately. Not
sure whether this is a recent change on their end or if it's always
been like this, because we didn't have test cases for sets with more
than 5 tracks.