refactor(ecau): standardise 404 error messages for providers
Previously, when a release returned a 404 code, a generic HTTP error
would be shown in the log. Changing this to be more specific and to
include the provider name in the error message.

Not all providers have been changed though. Notably the Tidal provider
will still throw a generic error because it performs custom requests.
That remains to be fixed.