fix(ecau): don't extract URLs from anchors if there are images
It's entirely possible for someone to copy-and-paste a whole chunk
of a document that includes both images and anchors. We probably
don't want to extract from the anchors if there are images. So we'll
first attempt to extract images, and fall back on anchors if there
are no images. If there are neither anchors nor images, we'll fall
back on generic parsing of plain text.