ci(deploy): check for changes against base instead of previous deploy
The continuous deployment checks whether changes have been made to the
scripts and deploys an update if so. However, it was checking against
the latest deployed version, rather than the "previous version". These
two are different if the `skip cd` tag is used. This led the deployment
to deploy updates for PRs that did not change the scripts at all, because
an old, not-yet-deployed change was skipped previously.

Instead, we'll now check against the compiled scripts based on the
"previous" commit ref. For pushes to main, this is the commit before
the push (even when multiple commits are pushed); for deployment
previews in PRs, this is the base branch of the PR; for checking
for changes in automated dependency updates, this is the main
branch.