ci: put npm cache on a monthly diet
Since we started using caching, the cache has grown in size 10-fold.
Once we switched CI to use Node 18, all of a sudden cache size dropped
substantially. The reason is that the npm cache is incremental, and it
doesn't remove package downloads that we're no longer using. The cache
therefore contains old package downloads that we haven't used in months,
and that's slowing down CI since it takes longer to fetch and extract
that cache.

Instead, we'll change the cache keys every month by having the keys
depend on the current year and month. Therefore, we'll start off with
a fresh cache on a monthly basis, and the cache won't waste unnecessary
space any longer.