style: disable new ESLint unicorn plugin rule
I don't fully agree with this, since IMO it makes more sense to
put the one-line exceptional conditions first rather than at the
very end. On the other hand, some of the suggestions are good.
Perhaps we could re-enable it if we refactor the problem areas a
bit to instead use early returns/continues for the exception
conditions.