ci(renovate): print correct script diffs
The diffs that were printed when a renovate update led to a change
in the built scripts, was compared against the last deployed version,
not the last commit. Therefore, it includes any change that we previously
skipped deploying, making it difficult to determine the changes that the
current upgrade causes.