ci: configure renovate to create a single PR per month
The package update spam is getting out of hand, and the vast majority
of recent commits are updates to dev dependencies. The previous config
could make sense when the repo is very active, but this repo hasn't been
very active in the past few months except for merging dependency updates.
That maintenance effort is better spent elsewhere, so reducing the frequency
of automated updates makes sense.

Patch and minor upgrades should all be grouped into a single monthly PR,
and will be auto-merged if CI passes and the compiled scripts are
unaffected. Major upgrades and any upgrade to production dependencies
will still be raised as a PR, but also just once a month.

We're explicitly ignoring one package due to #624, so if a future
update to that package is available, it'll still be a separate PR.
If that update is merged, we should also remove it from the ignored
packages.