test: fix crypto.subtle.digest errors on newer Node versions
Seems like a recent Node version has started to do stricter checks
on the data passed into the webcrypto APIs. In lib/util/blob.ts, we're
passing in the result of a FileReader, which is polyfilled by jest and
is using a fake ArrayBuffer, which fails this check. Therefore, we need
to convert this fake ArrayBuffer into a valid Node object before passing
it to the API.