fix(ecau): prevent image blob from unloading before upload
It appears that Violentmonkey's implementation of `GM.xhr` with a blob
`responseType` causes the blob to become unloaded if the response itself
goes out of scope. When that happens, the image we enqueued for uploading
will become empty, and the upload will fail. To prevent this from happening,
we'll instead copy the buffer, which should prevent the contents from being
unloaded.