fix(ecau): automatically extract and refresh Soundcloud client IDs
Soundcloud has changed the client ID again, so instead of manually
updating it all of the time, we'll extract it. The extraction process
is the same as the one yt-dlp has been using for 3 years. This will
also automatically refresh the ID on the first auth failure, and
throw an error if it's still incorrect.

Tests are missing some edge cases that are difficult to cover, let's
ignore that for now.