test: fix `structuredClone` not being available in `IndexedDB` tests
`fake-indexeddb` relies on the `structuredClone` global in Node. However,
`jest-environment-jsdom` does not contain this function, so we need to
patch it into the environment.

Moreover, because of a bug in the JSDOM environment, we need to use
`toEqual` instead of `toStrictEqual` when checking the result of the
cache lookups from the `IndexedDB`.