fix(ecau): copy MusicBrainz complex artwork types
"Matrix/Runout" and "Raw/Unedited" could not be copied when a MB
release's images were copied because the enum used the "Matrix" and
"Raw" keys only. Instead, we'll use the full artwork type name as
keys. This also requires changes in VGMdb's caption to artwork type
mappings and related tests. Also add test case with a real release.