fix: increase SPA ID length limit for work code validation
SPA IDs are sequential numbers. Sometime within the last 2.5 years, the 10 millionth SPA ID was issued, which means that an increasing amount of IDs are now 8 characters long instead of 7. However, the script associated with this library only looks for 7 digits, and presents (incorrect) validation errors for 8-digit SPA IDs.