style: properly enable JSON linting
It was not enabled before, because `.json` wasn't considered as a
file extension. However, adding it as a file extension causes issues
with the `@typescript-eslint` plugin, so we need to separate the
configuration for JSON files from the rest of the JS rules. The new
ESLint flat configuration will allow us to be more elegant about this
soon.