style: always return in promise handlers
We had disabled this rule, but I have now changed my mind since we can
ignore promise handlers that are the last part of the chain.

Changed the `logFailure` utility function so that it doesn't
take a promise but instead returns a catch handler. Otherwise, since
the promise is provided as an argument, a warning would still be emitted
for missing returns.