fix(ecau): fix Audible audiobook cover extraction (again)
We now use the same extraction logic as the other pages, so the
implementation is also severely simplified.

Also update another test case since the upstream page has changed.
We're changing it to a different release because we need to test
whether we can extract multiple images too, and the new upstream
only has one.