fix(ecau): extract RateYourMusic covers again
The full-resolution covers are now locked behind a Captcha, but we can
get a lower-resolution 1200px cover from the standard pages.

We're also adding a new build script, `replace-recording`, that can take
a HAR file created elsewhere and inject its content into an existing WARC
recording to replace it. This is useful for testing certain pages whose
responses won't load properly outside of the browser (e.g., because of
Cloudflare protection). This makes it possible to create a recording
in the browser, export it to a HAR file, and inject its content in the
test WARC files. For instance, in Firefox, using the browser toolbox,
it is possible to export the network request made by the userscript
so that it can be used in the tests.

Finally, this also adds a check for Cloudflare captcha pages to display
a custom error message in case it appears, as may be the case for RYM.