fix(ecau): use larger thumbnails in Bandcamp deduplication
Deduplicating track images based on the 150x150 thumbnails appears
to be unreliable: The tests now failed, even though the source images
are identical. Ideally, we'd deduplicate based on full-size images, but
that'll be slow. Perhaps in the future we could load only the first few
kilobytes of the original images and make a decision based on that.