test(ecau): fix Discogs tests
They were disabled a while ago because of Cloudflare protection.
Now that we have the recording replacer script, we can actually update
the recordings and start using the tests again.

Also updated the negative test case as it pointed to a release that
now exists. The new ID is the largest ID that the API currently accepts.