chore(build): re-enable renovate on-demand
Renovate is scheduled to only run on Feb 29th, which is once every
4 years. I'm hoping that this will have the following effect:
- Renovate will not create update commits or PRs except one day of the
  year, thereby reducing its spam.
- Renovate will still monitor for updates and maintain its dependency
  dashboard, where the available updates will be marked as "awaiting
  schedule" and PRs can be created on demand.

Effectively, if this works, we'll be able to have an overview of
outdated dependencies, while also being able to manually curate
which pending updates we want to apply, all without a constant
barrage of automated commits on a dormant repository, or a constant
barrage of PR update e-mails. As far as I know, there is no way to
configure Renovate in such a way.

Also removing the grouping of dev dependencies. This was a good
strategy to reduce spammy commits, but if one of the dev dependencies
causes a breaking change, all updates will be blocked. If we'll
manually pick and choose the updates, it's less necessary. Ideally they'd
still be somewhat grouped (with separate groups for ESLint, TypeScript,
Jest, and rollup) but let's not over-engineer it just yet.