perf(ecau): optimise track image deduplication
First compare images according to their file size. If the file size
differs, then there's no point in comparing their content as it
will be different too.

Note that it is possible that file sizes cannot be extracted, due
to the necessary response headers being missing. When this happens,
we abort any comparisons based on file sizes, as they could be wrong.
This happens primarily with Soundcloud, which doesn't return the
content-length header.

According to the test recordings, this _does_ save some bandwidth
for Bandcamp releases by avoiding loading images unnecessarily.