fix(ecau): gracefully handle failures in Bandcamp track image extraction
The embedded player may be unavailable, e.g., on subscriber-only releases.
We can still load some cover art from those, so we'll avoid erroring out
when that happens.

The track images could technically be added using the old approach of
loading the track pages directly, but since I expect the combination of
subscriber-only releases AND track-specific images to be rare, I'm opting
to omit this functionality for now.

Note: Not sure why the Soundcloud recordings changed.